Estater_v1/requirements.txt
flask
gunicorn
tensorflow
numpy
Pillow
scikit-learn
